+++ app/Filament/Clusters/ProgramPpks.php
@@ -9,4 +9,5 @@ use Filament\Clusters\Cluster;
 class ProgramPpks extends Cluster
 {
     protected static ?string $navigationIcon = null;
+    protected static ?string $navigationLabel = 'Program PPKS';
 }
