+++ app/Filament/Clusters/ProgramRastra.php
@@ -9,4 +9,5 @@ use Filament\Clusters\Cluster;
 class ProgramRastra extends Cluster
 {
     protected static ?string $navigationIcon = null;
+    protected static ?string $navigationLabel = 'Program RASTRA';
 }
